Enable cross builds in jenkins with latest clang.
We have nightly clang tip of tree builds. They can be triggered in jenkins by
passing clang_gcc`some_version` (eg. clang_gcc52). Jenkins picks up latest
clang build and configures the environment with the corresponding version of the
gcc libstdc++.

This saves a lot of cpu cycles on the cxxmodules builds because we will reuse
a single set of binaries and stop building clang and llvm on every build node.