Correctly specify python2 for getPlatform.py
Right now we fail to source setup.sh here when python 3 is the default for the system. This for example happens with Guilherme's CVMFS Gentoo installation (that sadly has the only working gdb instance that one can use). So we explicitly have to state that this is python2 code that doesn't work with python3.