change enable_uring to ignore_uring for non-fedora machines
Instead of only running tests if the ROOTTEST_ENABLE_URING env var
exists, we define an env var ROOTTEST_IGNORE_URING that skips the
uring tests. This lets us easily run unit tests on non-fedora32 architectures
that support uring without breaking roottest for everyone else.

In the CMake test file we now check:

```
if(uring AND NOT DEFINED ENV{ROOTTEST_IGNORE_URING})
  ROOT_ADD_GTEST(RIoUring RIoUring.cxx LIBRARIES RIO uring)
endif()
```

which only succeeds if the uring is turned ON at build time and
the ROOTTEST_IGNORE_URING variable is not defined.